Refactor structure of project
