+++ lib/X1/Parser/Number.hs
@@ -5,7 +5,7 @@ import Protolude
 import qualified Data.Text as T
 import qualified Data.Text.Read as TR
 import qualified Data.Vector.Unboxed as V
-import X1.Parser.Types.Number
+import X1.Types.Expr1.Number
 import X1.Parser.Helpers
 import GHC.Unicode ( isDigit )
