+++ lib/X1/Parser/Pred.hs
@@ -3,8 +3,8 @@ module X1.Parser.Pred ( parser ) where
 import Protolude hiding ( Type )
 import X1.Parser.Helpers
 import qualified X1.Parser.Tyvar as Tyvar
-import X1.Parser.Types.Pred
-import X1.Parser.Types.Type
+import X1.Types.Expr1.Pred
+import X1.Types.Expr1.Type
 import X1.Types.Id
