+++ lib/X1/Parser/String.hs
@@ -2,7 +2,7 @@
 module X1.Parser.String ( parser ) where
 
 import Protolude
-import X1.Parser.Types.String
+import X1.Types.Expr1.String
 import X1.Parser.Helpers
 import Text.Megaparsec (takeWhileP)
