+++ lib/X1/Parser/Tycon.hs
@@ -2,7 +2,7 @@ module X1.Parser.Tycon ( parser ) where
 
 import Protolude
 import X1.Parser.Helpers
-import X1.Parser.Types.Tycon
+import X1.Types.Expr1.Tycon
 import X1.Types.Id
