+++ lib/X1/Parser/Type.hs
@@ -4,7 +4,7 @@ module X1.Parser.Type ( parser ) where
 import Protolude hiding ( Type, try )
 import qualified X1.Parser.Tycon as Tycon
 import qualified X1.Parser.Tyvar as Tyvar
-import X1.Parser.Types.Type
+import X1.Types.Expr1.Type
 import X1.Parser.Helpers
 import X1.Types.Id
 import Control.Monad.Combinators.Expr
