+++ lib/X1/Parser/Tyvar.hs
@@ -2,7 +2,7 @@ module X1.Parser.Tyvar ( parser ) where
 
 import Protolude
 import X1.Parser.Helpers
-import X1.Parser.Types.Tyvar
+import X1.Types.Expr1.Tyvar
 import X1.Types.Id
