+++ lib/X1/Types/Expr1/Pred.hs
@@ -1,8 +1,8 @@
 
-module X1.Parser.Types.Pred ( Pred(..) ) where
+module X1.Types.Expr1.Pred ( Pred(..) ) where
 
 import Protolude hiding ( Type )
-import X1.Parser.Types.Type
+import X1.Types.Expr1.Type
 import X1.Types.Id
