+++ tests/Test/X1/Parser/Expr1.hs
@@ -4,15 +4,15 @@ module Test.X1.Parser.Expr1 ( module Test.X1.Parser.Expr1 ) where
 import Protolude hiding ( Type )
 import Test.Tasty.Hspec
 import Test.X1.Parser.Helpers
-import X1.Types.Expr1
-import X1.Types.Pattern
-import X1.Types.Lit
-import X1.Types.Id
 import X1.Parser.Expr1 (parser)
-import X1.Parser.Types.String
-import X1.Parser.Types.Type
-import X1.Parser.Types.Scheme
-import X1.Parser.Types.Number
+import X1.Types.Id
+import X1.Types.Expr1
+import X1.Types.Expr1.Lit
+import X1.Types.Expr1.Type
+import X1.Types.Expr1.String
+import X1.Types.Expr1.Scheme
+import X1.Types.Expr1.Number
+import X1.Types.Expr1.Pattern
 import Test.Hspec.Megaparsec hiding (shouldFailWith, succeedsLeaving)
